Add files via upload
This project was created as a project in one of the courses named as Fundamentals of Computer Programming (FCP).
It is just a program in C language which is a basic implementation of the file management. It allows the user to just add, update, delete and search for a particular medicine in his/her own store as well as allows customers to search a particular medicine in 10 stores. The database is just an arbitrary database downloaded from the internet and not a real database.